mqtt-project-rasp/service-worker/requirements.txt
celery==5.2.3
redis==4.2.2
paho-mqtt==1.6.1
SQLAlchemy==1.4.31
PyMySQL==1.0.2
cryptography==3.4.7
